Nodo_1/requirements.txt
uvicorn
fastapi
jose
python-jose[cryptography]
pickle4
PyJWT
